backend/requirements.txt
fastapi
uvicorn
python-multipart
boto3
requests
python-dotenv
